@@ -9,42 +9,36 @@
     "deploy": "touch build/.nojekyll && npx gh-pages -d build -t true",
     "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-    "lint": "prettier --plugin-search-dir . --check . && eslint .",
-    "format": "prettier --plugin-search-dir . --write ."
+    "lint": "prettier --check . && eslint .",
+    "format": "prettier --write ."
   },
   "devDependencies": {
-    "@skeletonlabs/skeleton": "^1.0.0",
     "@sveltejs/adapter-auto": "^2.0.0",
-    "@sveltejs/adapter-static": "^2.0.1",
-    "@sveltejs/kit": "^1.5.0",
-    "@tailwindcss/forms": "^0.5.3",
-    "@tailwindcss/line-clamp": "^0.4.2",
-    "@tailwindcss/typography": "^0.5.9",
-    "@typescript-eslint/eslint-plugin": "^5.45.0",
-    "@typescript-eslint/parser": "^5.45.0",
+    "@sveltejs/adapter-static": "^2.0.3",
+    "@sveltejs/kit": "^1.27.4",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
     "autoprefixer": "^10.4.7",
     "eslint": "^8.28.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-svelte3": "^4.0.0",
-    "postcss": "^8.4.14",
-    "postcss-load-config": "^4.0.1",
-    "prettier": "^2.8.0",
-    "prettier-plugin-svelte": "^2.8.1",
-    "svelte": "^3.54.0",
-    "svelte-check": "^3.0.1",
-    "svelte-preprocess": "^4.10.7",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-svelte": "^2.30.0",
+    "postcss": "^8.4.31",
+    "prettier": "^3.0.0",
+    "prettier-plugin-svelte": "^3.0.0",
+    "svelte": "^4.2.7",
+    "svelte-check": "^3.6.0",
     "svelte-virtual-scroll-list": "^1.1.0",
-    "tailwindcss": "^3.1.5",
+    "tailwindcss": "^3.3.5",
     "tslib": "^2.4.1",
     "tw-colors": "^3.3.1",
-    "typescript": "^4.9.3",
+    "typescript": "^5.0.0",
     "vite": "^4.0.0"
   },
   "type": "module",
   "dependencies": {
     "@fontsource-variable/fira-code": "^5.0.15",
     "color-delta-e": "^1.1.2",
-    "iconify-icon": "^1.0.7",
+    "iconify-icon": "^1.0.8",
     "tailwind-merge": "^2.1.0"
   }
 }
